refactor so that using the PD API avoids unecessary imports
PD API no longer imports Kubernetes and the CRD.

In the create-statefulset-immediately PR this reduced the binary
by ~20M
