+++ pkg/apis/httputil/httputil.go
@@ -1,0 +1,44 @@
+package httputil
+
+import (
+	"fmt"
+	"io"
+	"io/ioutil"
+	"net/http"
+)
+
+// DeferClose captures the error returned from closing (if an error occurs).
+// This is designed to be used in a defer statement.
+func DeferClose(c io.Closer, err *error) {
+	if cerr := c.Close(); cerr != nil && *err == nil {
+		*err = cerr
+	}
+}
+
+// ReadErrorBody in the error case ready the body message.
+// But return it as an error (or return an error from reading the body).
+func ReadErrorBody(body io.Reader) (err error) {
+	bodyBytes, err := ioutil.ReadAll(body)
+	if err != nil {
+		return err
+	}
+	return fmt.Errorf(string(bodyBytes))
+}
+
+// GetBodyOK returns the body or an error if the response is not okay
+func GetBodyOK(httpClient *http.Client, apiURL string) ([]byte, error) {
+	res, err := httpClient.Get(apiURL)
+	if err != nil {
+		return nil, err
+	}
+	defer DeferClose(res.Body, &err)
+	if res.StatusCode >= 400 {
+		errMsg := fmt.Errorf(fmt.Sprintf("Error response %v URL %s", res.StatusCode, apiURL))
+		return nil, errMsg
+	}
+	body, err := ioutil.ReadAll(res.Body)
+	if err != nil {
+		return nil, err
+	}
+	return body, err
+}
