+++ pkg/controller/pod_control_test.go
@@ -21,6 +21,7 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
+	"github.com/pingcap/tidb-operator/pkg/apis/pdapi"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap.com/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/label"
 	corev1 "k8s.io/api/core/v1"
@@ -43,16 +44,15 @@ func TestPodControlUpdateMetaInfoSuccess(t *testing.T) {
 	pod := newPod(tc)
 	fakeClient, pdControl, podLister, _, recorder := newFakeClientRecorderAndPDControl()
 	control := NewRealPodControl(fakeClient, pdControl, podLister, recorder)
-	pdClient := NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
-	pdClient.AddReaction(GetClusterActionType, func(action *Action) (interface{}, error) {
+	pdClient := NewFakePDClient(pdControl, tc)
+	pdClient.AddReaction(pdapi.GetClusterActionType, func(action *pdapi.Action) (interface{}, error) {
 		cluster := &metapb.Cluster{
 			Id: 222,
 		}
 		return cluster, nil
 	})
-	pdClient.AddReaction(GetMembersActionType, func(action *Action) (interface{}, error) {
-		membersInfo := &MembersInfo{
+	pdClient.AddReaction(pdapi.GetMembersActionType, func(action *pdapi.Action) (interface{}, error) {
+		membersInfo := &pdapi.MembersInfo{
 			Members: []*pdpb.Member{
 				{
 					MemberId: 111,
@@ -61,11 +61,11 @@ func TestPodControlUpdateMetaInfoSuccess(t *testing.T) {
 		}
 		return membersInfo, nil
 	})
-	pdClient.AddReaction(GetStoresActionType, func(action *Action) (interface{}, error) {
-		storesInfo := &StoresInfo{
-			Stores: []*StoreInfo{
+	pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
+		storesInfo := &pdapi.StoresInfo{
+			Stores: []*pdapi.StoreInfo{
 				{
-					Store: &MetaStore{
+					Store: &pdapi.MetaStore{
 						Store: &metapb.Store{
 							Id:      333,
 							Address: fmt.Sprintf("%s.web", TestPodName),
@@ -94,13 +94,12 @@ func TestPodControlUpdateMetaInfoGetClusterFailed(t *testing.T) {
 	pod := newPod(tc)
 	fakeClient, pdControl, podLister, _, recorder := newFakeClientRecorderAndPDControl()
 	control := NewRealPodControl(fakeClient, pdControl, podLister, recorder)
-	pdClient := NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
-	pdClient.AddReaction(GetClusterActionType, func(action *Action) (interface{}, error) {
+	pdClient := NewFakePDClient(pdControl, tc)
+	pdClient.AddReaction(pdapi.GetClusterActionType, func(action *pdapi.Action) (interface{}, error) {
 		return nil, errors.New("failed to get cluster info from PD server")
 	})
-	pdClient.AddReaction(GetMembersActionType, func(action *Action) (interface{}, error) {
-		membersInfo := &MembersInfo{
+	pdClient.AddReaction(pdapi.GetMembersActionType, func(action *pdapi.Action) (interface{}, error) {
+		membersInfo := &pdapi.MembersInfo{
 			Members: []*pdpb.Member{
 				{
 					MemberId: 111,
@@ -109,11 +108,11 @@ func TestPodControlUpdateMetaInfoGetClusterFailed(t *testing.T) {
 		}
 		return membersInfo, nil
 	})
-	pdClient.AddReaction(GetStoresActionType, func(action *Action) (interface{}, error) {
-		storesInfo := &StoresInfo{
-			Stores: []*StoreInfo{
+	pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
+		storesInfo := &pdapi.StoresInfo{
+			Stores: []*pdapi.StoreInfo{
 				{
-					Store: &MetaStore{
+					Store: &pdapi.MetaStore{
 						Store: &metapb.Store{
 							Id:      333,
 							Address: fmt.Sprintf("%s.web", TestPodName),
@@ -141,22 +140,21 @@ func TestPodControlUpdateMetaInfoGetMemberFailed(t *testing.T) {
 	pod := newPod(tc)
 	fakeClient, pdControl, podLister, _, recorder := newFakeClientRecorderAndPDControl()
 	control := NewRealPodControl(fakeClient, pdControl, podLister, recorder)
-	pdClient := NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
-	pdClient.AddReaction(GetClusterActionType, func(action *Action) (interface{}, error) {
+	pdClient := NewFakePDClient(pdControl, tc)
+	pdClient.AddReaction(pdapi.GetClusterActionType, func(action *pdapi.Action) (interface{}, error) {
 		cluster := &metapb.Cluster{
 			Id: 222,
 		}
 		return cluster, nil
 	})
-	pdClient.AddReaction(GetMembersActionType, func(action *Action) (interface{}, error) {
+	pdClient.AddReaction(pdapi.GetMembersActionType, func(action *pdapi.Action) (interface{}, error) {
 		return nil, errors.New("failed to get member info from PD server")
 	})
-	pdClient.AddReaction(GetStoresActionType, func(action *Action) (interface{}, error) {
-		storesInfo := &StoresInfo{
-			Stores: []*StoreInfo{
+	pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
+		storesInfo := &pdapi.StoresInfo{
+			Stores: []*pdapi.StoreInfo{
 				{
-					Store: &MetaStore{
+					Store: &pdapi.MetaStore{
 						Store: &metapb.Store{
 							Id:      333,
 							Address: fmt.Sprintf("%s.web", TestPodName),
@@ -185,16 +183,15 @@ func TestPodControlUpdateMetaInfoGetStoreFailed(t *testing.T) {
 	pod := newPod(tc)
 	fakeClient, pdControl, podLister, _, recorder := newFakeClientRecorderAndPDControl()
 	control := NewRealPodControl(fakeClient, pdControl, podLister, recorder)
-	pdClient := NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
-	pdClient.AddReaction(GetClusterActionType, func(action *Action) (interface{}, error) {
+	pdClient := NewFakePDClient(pdControl, tc)
+	pdClient.AddReaction(pdapi.GetClusterActionType, func(action *pdapi.Action) (interface{}, error) {
 		cluster := &metapb.Cluster{
 			Id: 222,
 		}
 		return cluster, nil
 	})
-	pdClient.AddReaction(GetMembersActionType, func(action *Action) (interface{}, error) {
-		membersInfo := &MembersInfo{
+	pdClient.AddReaction(pdapi.GetMembersActionType, func(action *pdapi.Action) (interface{}, error) {
+		membersInfo := &pdapi.MembersInfo{
 			Members: []*pdpb.Member{
 				{
 					MemberId: 111,
@@ -203,7 +200,7 @@ func TestPodControlUpdateMetaInfoGetStoreFailed(t *testing.T) {
 		}
 		return membersInfo, nil
 	})
-	pdClient.AddReaction(GetStoresActionType, func(action *Action) (interface{}, error) {
+	pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 		return nil, errors.New("failed to get store info from PD server")
 	})
 
@@ -224,16 +221,15 @@ func TestPodControlUpdateMetaInfoUpdatePodFailed(t *testing.T) {
 	pod := newPod(tc)
 	fakeClient, pdControl, podLister, _, recorder := newFakeClientRecorderAndPDControl()
 	control := NewRealPodControl(fakeClient, pdControl, podLister, recorder)
-	pdClient := NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
-	pdClient.AddReaction(GetClusterActionType, func(action *Action) (interface{}, error) {
+	pdClient := NewFakePDClient(pdControl, tc)
+	pdClient.AddReaction(pdapi.GetClusterActionType, func(action *pdapi.Action) (interface{}, error) {
 		cluster := &metapb.Cluster{
 			Id: 222,
 		}
 		return cluster, nil
 	})
-	pdClient.AddReaction(GetMembersActionType, func(action *Action) (interface{}, error) {
-		membersInfo := &MembersInfo{
+	pdClient.AddReaction(pdapi.GetMembersActionType, func(action *pdapi.Action) (interface{}, error) {
+		membersInfo := &pdapi.MembersInfo{
 			Members: []*pdpb.Member{
 				{
 					MemberId: 111,
@@ -242,11 +238,11 @@ func TestPodControlUpdateMetaInfoUpdatePodFailed(t *testing.T) {
 		}
 		return membersInfo, nil
 	})
-	pdClient.AddReaction(GetStoresActionType, func(action *Action) (interface{}, error) {
-		storesInfo := &StoresInfo{
-			Stores: []*StoreInfo{
+	pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
+		storesInfo := &pdapi.StoresInfo{
+			Stores: []*pdapi.StoreInfo{
 				{
-					Store: &MetaStore{
+					Store: &pdapi.MetaStore{
 						Store: &metapb.Store{
 							Id:      333,
 							Address: fmt.Sprintf("%s.web", TestPodName),
@@ -278,16 +274,15 @@ func TestPodControlUpdateMetaInfoConflictSuccess(t *testing.T) {
 	fakeClient, pdControl, podLister, podIndexer, recorder := newFakeClientRecorderAndPDControl()
 	podIndexer.Add(oldPod)
 	control := NewRealPodControl(fakeClient, pdControl, podLister, recorder)
-	pdClient := NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
-	pdClient.AddReaction(GetClusterActionType, func(action *Action) (interface{}, error) {
+	pdClient := NewFakePDClient(pdControl, tc)
+	pdClient.AddReaction(pdapi.GetClusterActionType, func(action *pdapi.Action) (interface{}, error) {
 		cluster := &metapb.Cluster{
 			Id: 222,
 		}
 		return cluster, nil
 	})
-	pdClient.AddReaction(GetMembersActionType, func(action *Action) (interface{}, error) {
-		membersInfo := &MembersInfo{
+	pdClient.AddReaction(pdapi.GetMembersActionType, func(action *pdapi.Action) (interface{}, error) {
+		membersInfo := &pdapi.MembersInfo{
 			Members: []*pdpb.Member{
 				{
 					MemberId: 111,
@@ -296,11 +291,11 @@ func TestPodControlUpdateMetaInfoConflictSuccess(t *testing.T) {
 		}
 		return membersInfo, nil
 	})
-	pdClient.AddReaction(GetStoresActionType, func(action *Action) (interface{}, error) {
-		storesInfo := &StoresInfo{
-			Stores: []*StoreInfo{
+	pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
+		storesInfo := &pdapi.StoresInfo{
+			Stores: []*pdapi.StoreInfo{
 				{
-					Store: &MetaStore{
+					Store: &pdapi.MetaStore{
 						Store: &metapb.Store{
 							Id:      333,
 							Address: fmt.Sprintf("%s.web", TestPodName),
@@ -385,9 +380,9 @@ func TestPodControlUpdatePodConflictSuccess(t *testing.T) {
 	g.Expect(events[0]).To(ContainSubstring(corev1.EventTypeNormal))
 }
 
-func newFakeClientRecorderAndPDControl() (*fake.Clientset, *FakePDControl, corelisters.PodLister, cache.Indexer, *record.FakeRecorder) {
+func newFakeClientRecorderAndPDControl() (*fake.Clientset, *pdapi.FakePDControl, corelisters.PodLister, cache.Indexer, *record.FakeRecorder) {
 	fakeClient := &fake.Clientset{}
-	pdControl := NewFakePDControl()
+	pdControl := pdapi.NewFakePDControl()
 	kubeCli := kubefake.NewSimpleClientset()
 	recorder := record.NewFakeRecorder(10)
 	kubeInformerFactory := kubeinformers.NewSharedInformerFactory(kubeCli, 0)
