+++ pkg/manager/member/pd_upgrader_test.go
@@ -18,6 +18,7 @@ import (
 	"testing"
 
 	. "github.com/onsi/gomega"
+	"github.com/pingcap/tidb-operator/pkg/apis/pdapi"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap.com/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/controller"
 	"github.com/pingcap/tidb-operator/pkg/label"
@@ -46,20 +47,19 @@ func TestPDUpgraderUpgrade(t *testing.T) {
 	testFn := func(test *testcase) {
 		t.Log(test.name)
 		upgrader, pdControl, _, podInformer := newPDUpgrader()
-		pdClient := controller.NewFakePDClient()
 		tc := newTidbClusterForPDUpgrader()
-		pdControl.SetPDClient(tc, pdClient)
+		pdClient := controller.NewFakePDClient(pdControl, tc)
 
 		if test.changeFn != nil {
 			test.changeFn(tc)
 		}
 
 		if test.transferLeaderErr {
-			pdClient.AddReaction(controller.TransferPDLeaderActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.TransferPDLeaderActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, fmt.Errorf("failed to transfer leader")
 			})
 		} else {
-			pdClient.AddReaction(controller.TransferPDLeaderActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.TransferPDLeaderActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, nil
 			})
 		}
@@ -221,10 +221,10 @@ func TestPDUpgraderUpgrade(t *testing.T) {
 
 }
 
-func newPDUpgrader() (Upgrader, *controller.FakePDControl, *controller.FakePodControl, podinformers.PodInformer) {
+func newPDUpgrader() (Upgrader, *pdapi.FakePDControl, *controller.FakePodControl, podinformers.PodInformer) {
 	kubeCli := kubefake.NewSimpleClientset()
 	podInformer := kubeinformers.NewSharedInformerFactory(kubeCli, 0).Core().V1().Pods()
-	pdControl := controller.NewFakePDControl()
+	pdControl := pdapi.NewFakePDControl()
 	podControl := controller.NewFakePodControl(podInformer)
 	return &pdUpgrader{
 			pdControl:  pdControl,
