+++ pkg/manager/member/tikv_member_manager_test.go
@@ -21,6 +21,7 @@ import (
 
 	. "github.com/onsi/gomega"
 	"github.com/pingcap/kvproto/pkg/metapb"
+	"github.com/pingcap/tidb-operator/pkg/apis/pdapi"
 	"github.com/pingcap/tidb-operator/pkg/apis/pingcap.com/v1alpha1"
 	"github.com/pingcap/tidb-operator/pkg/client/clientset/versioned/fake"
 	informers "github.com/pingcap/tidb-operator/pkg/client/informers/externalversions"
@@ -48,8 +49,8 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 		err                          bool
 		tikvPeerSvcCreated           bool
 		setCreated                   bool
-		pdStores                     *controller.StoresInfo
-		tombstoneStores              *controller.StoresInfo
+		pdStores                     *pdapi.StoresInfo
+		tombstoneStores              *pdapi.StoresInfo
 	}
 
 	testFn := func(test *testcase, t *testing.T) {
@@ -73,17 +74,17 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 		tkmm, fakeSetControl, fakeSvcControl, pdClient, _, _ := newFakeTiKVMemberManager(tc)
 
 		if test.errWhenGetStores {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, fmt.Errorf("failed to get stores from tikv cluster")
 			})
 		} else {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.pdStores, nil
 			})
-			pdClient.AddReaction(controller.GetTombStoneStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetTombStoneStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.tombstoneStores, nil
 			})
-			pdClient.AddReaction(controller.SetStoreLabelsActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.SetStoreLabelsActionType, func(action *pdapi.Action) (interface{}, error) {
 				return true, nil
 			})
 		}
@@ -130,8 +131,8 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 			err:                          false,
 			tikvPeerSvcCreated:           true,
 			setCreated:                   true,
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 		},
 		{
 			name: "pd is not available",
@@ -143,8 +144,8 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 			err:                          true,
 			tikvPeerSvcCreated:           false,
 			setCreated:                   false,
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 		},
 		{
 			name: "tidbcluster's storage format is wrong",
@@ -156,8 +157,8 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 			err:                          true,
 			tikvPeerSvcCreated:           true,
 			setCreated:                   false,
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 		},
 		{
 			name:                         "error when create statefulset",
@@ -167,8 +168,8 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 			err:                          true,
 			tikvPeerSvcCreated:           true,
 			setCreated:                   false,
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 		},
 		{
 			name:                         "error when create tikv peer service",
@@ -178,8 +179,8 @@ func TestTiKVMemberManagerSyncCreate(t *testing.T) {
 			err:                          true,
 			tikvPeerSvcCreated:           false,
 			setCreated:                   false,
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 		},
 	}
 
@@ -193,8 +194,8 @@ func TestTiKVMemberManagerSyncUpdate(t *testing.T) {
 	type testcase struct {
 		name                         string
 		modify                       func(cluster *v1alpha1.TidbCluster)
-		pdStores                     *controller.StoresInfo
-		tombstoneStores              *controller.StoresInfo
+		pdStores                     *pdapi.StoresInfo
+		tombstoneStores              *pdapi.StoresInfo
 		errWhenUpdateStatefulSet     bool
 		errWhenUpdateTiKVPeerService bool
 		errWhenGetStores             bool
@@ -221,17 +222,17 @@ func TestTiKVMemberManagerSyncUpdate(t *testing.T) {
 
 		tkmm, fakeSetControl, fakeSvcControl, pdClient, _, _ := newFakeTiKVMemberManager(tc)
 		if test.errWhenGetStores {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, fmt.Errorf("failed to get stores from pd cluster")
 			})
 		} else {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.pdStores, nil
 			})
-			pdClient.AddReaction(controller.GetTombStoneStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetTombStoneStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.tombstoneStores, nil
 			})
-			pdClient.AddReaction(controller.SetStoreLabelsActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.SetStoreLabelsActionType, func(action *pdapi.Action) (interface{}, error) {
 				return true, nil
 			})
 		}
@@ -297,8 +298,8 @@ func TestTiKVMemberManagerSyncUpdate(t *testing.T) {
 				tc.Status.PD.Phase = v1alpha1.NormalPhase
 			},
 			// TODO add unit test for status sync
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 			errWhenUpdateStatefulSet:     false,
 			errWhenUpdateTiKVPeerService: false,
 			errWhenGetStores:             false,
@@ -320,8 +321,8 @@ func TestTiKVMemberManagerSyncUpdate(t *testing.T) {
 				tc.Spec.TiKV.Requests.Storage = "100xxxxi"
 				tc.Status.PD.Phase = v1alpha1.NormalPhase
 			},
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 			errWhenUpdateStatefulSet:     false,
 			errWhenUpdateTiKVPeerService: false,
 			err:                          true,
@@ -334,8 +335,8 @@ func TestTiKVMemberManagerSyncUpdate(t *testing.T) {
 				tc.Spec.TiKV.Replicas = 5
 				tc.Status.PD.Phase = v1alpha1.NormalPhase
 			},
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 			errWhenUpdateStatefulSet:     true,
 			errWhenUpdateTiKVPeerService: false,
 			err:                          true,
@@ -350,8 +351,8 @@ func TestTiKVMemberManagerSyncUpdate(t *testing.T) {
 				tc.Spec.TiKV.Replicas = 5
 				tc.Status.PD.Phase = v1alpha1.NormalPhase
 			},
-			pdStores:                     &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
-			tombstoneStores:              &controller.StoresInfo{Count: 0, Stores: []*controller.StoreInfo{}},
+			pdStores:                     &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
+			tombstoneStores:              &pdapi.StoresInfo{Count: 0, Stores: []*pdapi.StoreInfo{}},
 			errWhenUpdateStatefulSet:     false,
 			errWhenUpdateTiKVPeerService: false,
 			errWhenGetStores:             true,
@@ -481,7 +482,7 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		errWhenGetStores bool
 		hasNode          bool
 		hasPod           bool
-		storeInfo        *controller.StoresInfo
+		storeInfo        *pdapi.StoresInfo
 		errExpectFn      func(*GomegaWithT, error)
 		setCount         int
 		labelSetFailed   bool
@@ -491,12 +492,12 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		pmm, _, _, pdClient, podIndexer, nodeIndexer := newFakeTiKVMemberManager(tc)
 
 		if test.errWhenGetStores {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, fmt.Errorf("failed to get stores")
 			})
 		}
 		if test.storeInfo != nil {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.storeInfo, nil
 			})
 		}
@@ -527,11 +528,11 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 			podIndexer.Add(pod)
 		}
 		if test.labelSetFailed {
-			pdClient.AddReaction(controller.SetStoreLabelsActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.SetStoreLabelsActionType, func(action *pdapi.Action) (interface{}, error) {
 				return false, fmt.Errorf("label set failed")
 			})
 		} else {
-			pdClient.AddReaction(controller.SetStoreLabelsActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.SetStoreLabelsActionType, func(action *pdapi.Action) (interface{}, error) {
 				return true, nil
 			})
 		}
@@ -559,8 +560,8 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "stores is empty",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			hasNode: true,
 			hasPod:  true,
@@ -573,8 +574,8 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "status is nil",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
 						Status: nil,
 					},
@@ -591,8 +592,8 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "store is nil",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
 						Store: nil,
 					},
@@ -609,17 +610,17 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "don't have pod",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LeaderCount:     1,
 							LastHeartbeatTS: time.Now(),
 						},
@@ -638,17 +639,17 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "don't have node",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LeaderCount:     1,
 							LastHeartbeatTS: time.Now(),
 						},
@@ -666,10 +667,10 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "already has labels",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
@@ -694,7 +695,7 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LeaderCount:     1,
 							LastHeartbeatTS: time.Now(),
 						},
@@ -712,10 +713,10 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "labels not equal, but set failed",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
@@ -728,7 +729,7 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LeaderCount:     1,
 							LastHeartbeatTS: time.Now(),
 						},
@@ -746,10 +747,10 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 		{
 			name:             "labels not equal, set success",
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
@@ -762,7 +763,7 @@ func TestTiKVMemberManagerSetStoreLabelsForTiKV(t *testing.T) {
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LeaderCount:     1,
 							LastHeartbeatTS: time.Now(),
 						},
@@ -790,11 +791,11 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 	type testcase struct {
 		name                      string
 		updateTC                  func(*v1alpha1.TidbCluster)
-		upgradingFn               func(corelisters.PodLister, controller.PDControlInterface, *apps.StatefulSet, *v1alpha1.TidbCluster) (bool, error)
+		upgradingFn               func(corelisters.PodLister, pdapi.PDControlInterface, *apps.StatefulSet, *v1alpha1.TidbCluster) (bool, error)
 		errWhenGetStores          bool
-		storeInfo                 *controller.StoresInfo
+		storeInfo                 *pdapi.StoresInfo
 		errWhenGetTombstoneStores bool
-		tombstoneStoreInfo        *controller.StoresInfo
+		tombstoneStoreInfo        *pdapi.StoresInfo
 		errExpectFn               func(*GomegaWithT, error)
 		tcExpectFn                func(*GomegaWithT, *v1alpha1.TidbCluster)
 	}
@@ -817,20 +818,20 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 			pmm.tikvStatefulSetIsUpgradingFn = test.upgradingFn
 		}
 		if test.errWhenGetStores {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, fmt.Errorf("failed to get stores")
 			})
 		} else if test.storeInfo != nil {
-			pdClient.AddReaction(controller.GetStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.storeInfo, nil
 			})
 		}
 		if test.errWhenGetTombstoneStores {
-			pdClient.AddReaction(controller.GetTombStoneStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetTombStoneStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return nil, fmt.Errorf("failed to get tombstone stores")
 			})
 		} else if test.tombstoneStoreInfo != nil {
-			pdClient.AddReaction(controller.GetTombStoneStoresActionType, func(action *controller.Action) (interface{}, error) {
+			pdClient.AddReaction(pdapi.GetTombStoneStoresActionType, func(action *pdapi.Action) (interface{}, error) {
 				return test.tombstoneStoreInfo, nil
 			})
 		}
@@ -847,7 +848,7 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "whether statefulset is upgrading returns failed",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, fmt.Errorf("whether upgrading failed")
 			},
 			errWhenGetStores:          false,
@@ -865,7 +866,7 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "statefulset is upgrading",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return true, nil
 			},
 			errWhenGetStores:          false,
@@ -883,7 +884,7 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 			updateTC: func(tc *v1alpha1.TidbCluster) {
 				tc.Status.PD.Phase = v1alpha1.UpgradePhase
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return true, nil
 			},
 			errWhenGetStores:          false,
@@ -899,7 +900,7 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "statefulset is not upgrading",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores:          false,
@@ -915,7 +916,7 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "get stores failed",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores:          true,
@@ -934,16 +935,16 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "stores is empty",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -955,20 +956,20 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "store is nil",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
 						Store: nil,
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -980,20 +981,20 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "status is nil",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
 						Status: nil,
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1010,28 +1011,28 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 					LastHeartbeatTime: now,
 				}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Time{},
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1050,28 +1051,28 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 					LastHeartbeatTime: metav1.Time{Time: time.Time{}},
 				}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Time{},
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1090,28 +1091,28 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 					LastHeartbeatTime: metav1.Time{Time: time.Time{}},
 				}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1128,28 +1129,28 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 				tc.Status.TiKV.Stores = map[string]v1alpha1.TiKVStore{}
 				// tc.Status.TiKV.Stores["333"] = v1alpha1.TiKVStore{}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1168,29 +1169,29 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 					State:              v1alpha1.TiKVStateUp,
 				}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1209,29 +1210,29 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 					State:              v1alpha1.TiKVStateUp,
 				}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Down",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: errExpectNil,
 			tcExpectFn: func(g *GomegaWithT, tc *v1alpha1.TidbCluster) {
@@ -1250,29 +1251,29 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 					State:              v1alpha1.TiKVStateUp,
 				}
 			},
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: true,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{},
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{},
 			},
 			errExpectFn: func(g *GomegaWithT, err error) {
 				g.Expect(err).To(HaveOccurred())
@@ -1287,38 +1288,38 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 		{
 			name:     "all works",
 			updateTC: nil,
-			upgradingFn: func(lister corelisters.PodLister, controlInterface controller.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
+			upgradingFn: func(lister corelisters.PodLister, controlInterface pdapi.PDControlInterface, set *apps.StatefulSet, cluster *v1alpha1.TidbCluster) (bool, error) {
 				return false, nil
 			},
 			errWhenGetStores: false,
-			storeInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			storeInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Up",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
 				},
 			},
 			errWhenGetTombstoneStores: false,
-			tombstoneStoreInfo: &controller.StoresInfo{
-				Stores: []*controller.StoreInfo{
+			tombstoneStoreInfo: &pdapi.StoresInfo{
+				Stores: []*pdapi.StoreInfo{
 					{
-						Store: &controller.MetaStore{
+						Store: &pdapi.MetaStore{
 							Store: &metapb.Store{
 								Id:      333,
 								Address: "pod-1.ns-1",
 							},
 							StateName: "Tombstone",
 						},
-						Status: &controller.StoreStatus{
+						Status: &pdapi.StoreStatus{
 							LastHeartbeatTS: time.Now(),
 						},
 					},
@@ -1341,12 +1342,11 @@ func TestTiKVMemberManagerSyncTidbClusterStatus(t *testing.T) {
 
 func newFakeTiKVMemberManager(tc *v1alpha1.TidbCluster) (
 	*tikvMemberManager, *controller.FakeStatefulSetControl,
-	*controller.FakeServiceControl, *controller.FakePDClient, cache.Indexer, cache.Indexer) {
+	*controller.FakeServiceControl, *pdapi.FakePDClient, cache.Indexer, cache.Indexer) {
 	cli := fake.NewSimpleClientset()
 	kubeCli := kubefake.NewSimpleClientset()
-	pdControl := controller.NewFakePDControl()
-	pdClient := controller.NewFakePDClient()
-	pdControl.SetPDClient(tc, pdClient)
+	pdControl := pdapi.NewFakePDControl()
+	pdClient := controller.NewFakePDClient(pdControl, tc)
 	setInformer := kubeinformers.NewSharedInformerFactory(kubeCli, 0).Apps().V1beta1().StatefulSets()
 	svcInformer := kubeinformers.NewSharedInformerFactory(kubeCli, 0).Core().V1().Services()
 	epsInformer := kubeinformers.NewSharedInformerFactory(kubeCli, 0).Core().V1().Endpoints()
